Add files via upload
When the rate of   infection is above metrics a warning and sound alert signal is thrown and records are no longer captured in the database. The medical officer would receive a warning text message on his or her mobile phone alerting users on the increase of the infection rate. These signals are used to inform the medical officers of the high rate of infection amongst students.